Update styling and functionality across application
Significant changes have been made to various aspects of the application's visual presentation and functionality, primarily focusing on typography size and color, as well as button styling. Moreover, new color themes have been added to TailwindCSS configuration file. The navigation bar was adjusted, with the addition of external links to social media platforms, and updates to the presentation of sections on the index page. Also, some components such as 'SIRCard', 'Section', and 'SirHero' were refined for consistency and improved UX.